+++ src/core/transformable.cpp
@@ -15,43 +15,43 @@ namespace pTK
     {
     }
     
-    const Vec2<uint32_t>& Transformable::get_size() const
+    const Vec2<float>& Transformable::get_size() const
     {
         return m_size;
     }
     
-    void Transformable::set_size(uint32_t width, uint32_t height)
+    void Transformable::set_size(float width, float height)
     {
         m_size = {width, height};
     }
     
-    void Transformable::set_size(const Vec2<uint32_t>& size)
+    void Transformable::set_size(const Vec2<float>& size)
     {
         m_size = size;
     }
     
-    const Vec2<uint32_t>& Transformable::get_position() const
+    const Vec2<float>& Transformable::get_position() const
     {
         return m_position;
     }
     
-    void Transformable::set_position(uint32_t x, uint32_t y)
+    void Transformable::set_position(float x, float y)
     {
         m_position = {x, y};
     }
     
-    void Transformable::set_position(const Vec2<uint32_t>& position)
+    void Transformable::set_position(const Vec2<float>& position)
     {
         m_position = position;
     }
     
-    void Transformable::move(uint32_t offset_x, uint32_t offset_y)
+    void Transformable::move(float offset_x, float offset_y)
     {
         m_position.x += offset_x;
         m_position.y += offset_y;
     }
     
-    void Transformable::move(const Vec2<uint32_t>& offset)
+    void Transformable::move(const Vec2<float>& offset)
     {
         m_position += offset;
     }
